+++ packages/react-paypal-js-storybook/v6/src/decorators/index.ts
@@ -1,1 +1,2 @@
 export { withPayPalProvider } from "./PayPalProviderDecorator";
+export { withPaymentResult } from "./PaymentResultDecorator";